+++ nevergrad/optimization/test_callbacks.py
@@ -19,10 +19,10 @@ def _func(x: Any, y: Any, blublu: str, array: Any) -> float:
 def test_log_parameters(tmp_path: Path) -> None:
     filepath = tmp_path / "logs.txt"
     cases = [0, np.int(1), np.float(2.0), np.nan, float("inf"), np.inf]
-    instrum = ng.Instrumentation(ng.p.Array(shape=(1,)),
-                                 ng.p.Scalar(),
-                                 blublu=ng.p.Choice(cases),
-                                 array=ng.p.Array(shape=(3, 2)))
+    instrum = ng.p.Instrumentation(ng.p.Array(shape=(1,)),
+                                   ng.p.Scalar(),
+                                   blublu=ng.p.Choice(cases),
+                                   array=ng.p.Array(shape=(3, 2)))
     optimizer = optimizerlib.OnePlusOne(instrumentation=instrum, budget=32)
     optimizer.register_callback("tell", callbacks.ParametersLogger(filepath, delete_existing_file=True))
     optimizer.minimize(_func, verbosity=2)
@@ -31,9 +31,9 @@ def test_log_parameters(tmp_path: Path) -> None:
     logs = logger.load_flattened()
     assert len(logs) == 32
     assert isinstance(logs[-1]["#arg1"], float)
-    assert len(logs[-1]) == 16
+    assert len(logs[-1]) == 18
     logs = logger.load_flattened(max_list_elements=2)
-    assert len(logs[-1]) == 12
+    assert len(logs[-1]) == 14
     # deletion
     logger = callbacks.ParametersLogger(filepath, delete_existing_file=True)
     assert not logger.load()