+++ nevergrad/optimization/test_optimizerlib.py
@@ -18,7 +18,6 @@ import pandas as pd
 from bayes_opt.util import acq_max
 from scipy import stats
 import nevergrad as ng
-from .. import instrumentation as inst
 from ..common.typetools import ArrayLike
 from ..common import testing
 from . import base
@@ -71,9 +70,9 @@ def check_optimizer(
             candidate = optimizer.minimize(fitness)
         if verify_value and "chain" not in str(optimizer_cls):
             try:
-                np.testing.assert_array_almost_equal(candidate.data, optimum, decimal=1)
+                np.testing.assert_array_almost_equal(candidate.args[0], optimum, decimal=1)
             except AssertionError as e:
-                print(f"Attemp #{k}: failed with best point {tuple(candidate.data)}")
+                print(f"Attemp #{k}: failed with best point {tuple(candidate.args[0])}")
                 if k == num_attempts:
                     raise e
             else:
@@ -92,7 +91,8 @@ def check_optimizer(
     # add a random point to test tell_not_asked
     assert not optimizer._asked, "All `ask`s  should have been followed by a `tell`"
     try:
-        candidate = optimizer.create_candidate.from_data(np.random.normal(0, 1, size=optimizer.dimension))
+        data = np.random.normal(0, 1, size=optimizer.dimension)
+        candidate = optimizer.instrumentation.spawn_child().set_standardized_data(data, deterministic=False)
         optimizer.tell(candidate, 12.0)
     except Exception as e:  # pylint: disable=broad-except
         if not isinstance(e, base.TellNotAskedNotSupportedError):
@@ -208,12 +208,12 @@ def test_optimizers_recommendation(with_parameter: bool,
         with patch("bayes_opt.bayesian_optimization.acq_max", patched):
             candidate = optim.minimize(fitness)
     if name not in recomkeeper.recommendations.index:
-        recomkeeper.recommendations.loc[name, :dimension] = tuple(candidate.data)
+        recomkeeper.recommendations.loc[name, :dimension] = tuple(candidate.args[0])
         raise ValueError(f'Recorded the value for optimizer "{name}", please rerun this test locally.')
     # BO slightly differs from a computer to another
     decimal = 2 if isinstance(optimizer_cls, optlib.ParametrizedBO) else 5
     np.testing.assert_array_almost_equal(
-        candidate.data,
+        candidate.args[0],
         recomkeeper.recommendations.loc[name, :][:dimension],
         decimal=decimal,
         err_msg="Something has changed, if this is normal, delete the following "
@@ -271,15 +271,16 @@ def test_tell_not_asked(name: str) -> None:
         opt.llambda = 2  # type: ignore
     else:
         opt._llambda = 2  # type: ignore
-    zeros = [0.0] * dim
-    opt.tell(opt.create_candidate.from_data(zeros), fitness(zeros))  # not asked
+    zero_c = opt.instrumentation.spawn_child().set_standardized_data([0.0] * dim)
+    best_c = opt.instrumentation.spawn_child().set_standardized_data(best)
+    opt.tell(zero_c, fitness(zero_c.args[0]))  # not asked
     asked = [opt.ask(), opt.ask()]
-    opt.tell(opt.create_candidate.from_data(best), fitness(best))  # not asked
+    opt.tell(best_c, fitness(best))  # not asked
     opt.tell(asked[0], fitness(*asked[0].args))
     opt.tell(asked[1], fitness(*asked[1].args))
     assert opt.num_tell == 4, opt.num_tell
     assert opt.num_ask == 2
-    if (0, 0, 0, 0) not in [tuple(x.data) for x in asked]:
+    if (0, 0, 0, 0) not in [tuple(x.args[0]) for x in asked]:
         for value in opt.archive.values():
             assert value.count == 1
 
@@ -292,7 +293,7 @@ def test_tbpsa_recom_with_update() -> None:
     optim.instrumentation.random_state.seed(12)
     optim.llambda = 3
     candidate = optim.minimize(fitness)
-    np.testing.assert_almost_equal(candidate.data, [0.037964, 0.0433031, -0.4688667, 0.3633273])
+    np.testing.assert_almost_equal(candidate.args[0], [0.037964, 0.0433031, -0.4688667, 0.3633273])
 
 
 def _square(x: np.ndarray, y: float = 12) -> float:
@@ -300,7 +301,7 @@ def _square(x: np.ndarray, y: float = 12) -> float:
 
 
 def test_optimization_doc_instrumentation_example() -> None:
-    instrum = inst.Instrumentation(inst.var.Array(2), y=inst.var.Array(1).asscalar())
+    instrum = ng.p.Instrumentation(ng.p.Array(shape=(2,)), y=ng.p.Scalar())
     optimizer = optlib.OnePlusOne(instrumentation=instrum, budget=100)
     recom = optimizer.minimize(_square)
     assert len(recom.args) == 1
@@ -329,7 +330,7 @@ def test_population_pickle(name: str) -> None:
 
 def test_bo_instrumentation_and_parameters() -> None:
     # instrumentation
-    instrumentation = inst.Instrumentation(ng.p.Choice([True, False]))
+    instrumentation = ng.p.Instrumentation(ng.p.Choice([True, False]))
     with pytest.warns(base.InefficientSettingsWarning):
         optlib.QRBO(instrumentation, budget=10)
     with pytest.warns(None) as record:
@@ -337,7 +338,8 @@ def test_bo_instrumentation_and_parameters() -> None:
     assert not record, record.list  # no warning
     # parameters
     # make sure underlying BO optimizer gets instantiated correctly
-    opt.tell(opt.create_candidate.from_call(True), 0.0)
+    new_candidate = opt.instrumentation.spawn_child(new_value=((True,), {}))
+    opt.tell(new_candidate, 0.0)
 
 
 def test_chaining() -> None:
@@ -354,7 +356,7 @@ def test_chaining() -> None:
 
 
 def test_instrumentation_optimizer_reproducibility() -> None:
-    instrumentation = inst.Instrumentation(ng.p.Array(shape=(1,)), y=ng.p.Choice(list(range(100))))
+    instrumentation = ng.p.Instrumentation(ng.p.Array(shape=(1,)), y=ng.p.Choice(list(range(100))))
     instrumentation.random_state.seed(12)
     optimizer = optlib.RandomSearch(instrumentation, budget=10)
     recom = optimizer.minimize(_square)